Ramdisk: Flash 1.0 profile
The goal of this profile is burst; as a result, we set input boost to 1248MHz for 50ms for initial ramp then use the target loads to aggressively use the little cluster. Due to the fact that the highest
frequency of the little cluster (1555MHz) uses less power than the stock minimum of the big cluster (633MHz), we can abuse the power of the little cluster and keep use of the big cluster down to a
minimum by keeping a high initial target load of 95; however, we will allow the big cluster to scale relatively easily up to 1248MHz if it really needs it (like with gaming or opening apps from scratch).
Overall, this profile should be good for various different uses as it will keep the phone bursty but use less energy than stock as the big cluster will be used less and the little cluster should not have
to do that much more work (as the stock value of input boost is 1344MHz for 40ms with target loads of 90 across all frequencies).

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>